Refactor: Extract sprite declarations to Drawing.Sprite.h
Move sprite related types and function declarations from Drawing.h, to a new Drawing.Sprite.h header. Also extract PaletteMap to its own header
